file renames, packages updated, variables added to resolvers and typeDefs
