--- client/src/App.css
@@ -1,244 +1,0 @@
-html {
-    line-height: 1.15;
-    /* 1 */
-    -webkit-text-size-adjust: 100%;
-    /* 2 */
-  }
-
-  body {
-    font-family: 'Rubik', sans-serif;
-  }
-  
-  h1,
-  h2,
-  h3,
-  h4,
-  h5,
-  h6 {
-    font-family: 'Rubik Mono One', sans-serif;
-  }
-  
-  .small {
-    font-weight: 300;
-    font-style: italic;
-  }
-  
-  .logo {
-    border-radius: 15px;
-    -webkit-animation: slit-in-diagonal-1 0.5s ease-out both, shadow-drop-center 2.5s cubic-bezier(0.250, 0.460, 0.450, 0.940) both;
-    animation: slit-in-diagonal-1 0.5s ease-out both, shadow-drop-center 2.5s cubic-bezier(0.250, 0.460, 0.450, 0.940) both;
-  }
-  
-  @-webkit-keyframes shadow-drop-center {
-    0% {
-      -webkit-box-shadow: 0 0 0 0 rgba(255, 255, 255, 0);
-      box-shadow: 0 0 0 0 rgba(255, 255, 255, 0);
-    }
-  
-    100% {
-      -webkit-box-shadow: 0 0 30px 10px rgba(255, 255, 255, 0.35);
-      box-shadow: 0 0 30px 10px rgba(255, 255, 255, 0.35);
-    }
-  }
-  
-  @keyframes shadow-drop-center {
-    0% {
-      -webkit-box-shadow: 0 0 0 0 rgba(255, 255, 255, 0);
-      box-shadow: 0 0 0 0 rgba(255, 255, 255, 0);
-    }
-  
-    100% {
-      -webkit-box-shadow: 0 0 30px 10px rgba(255, 255, 255, 0.35);
-      box-shadow: 0 0 30px 10px rgba(255, 255, 255, 0.35);
-    }
-  }
-  
-  
-  .mem-card:hover {
-    cursor: pointer;
-  }
-  
-  .mem-card {
-    -webkit-animation: slit-in-diagonal-1 0.6s ease-out both;
-    animation: slit-in-diagonal-1 0.6s ease-out both;
-  }
-  
-  @-webkit-keyframes slit-in-diagonal-1 {
-    0% {
-      -webkit-transform: translateZ(-800px) rotate3d(1, 1, 0, 90deg);
-      transform: translateZ(-800px) rotate3d(1, 1, 0, 90deg);
-      -webkit-animation-timing-function: ease-in;
-      animation-timing-function: ease-in;
-      opacity: 0;
-    }
-  
-    54% {
-      -webkit-transform: translateZ(-160px) rotate3d(1, 1, 0, 87deg);
-      transform: translateZ(-160px) rotate3d(1, 1, 0, 87deg);
-      -webkit-animation-timing-function: ease-in-out;
-      animation-timing-function: ease-in-out;
-      opacity: 1;
-    }
-  
-    100% {
-      -webkit-transform: translateZ(0) rotate3d(1, 1, 0, 0);
-      transform: translateZ(0) rotate3d(1, 1, 0, 0);
-      -webkit-animation-timing-function: ease-out;
-      animation-timing-function: ease-out;
-    }
-  }
-  
-  @keyframes slit-in-diagonal-1 {
-    0% {
-      -webkit-transform: translateZ(-800px) rotate3d(1, 1, 0, 90deg);
-      transform: translateZ(-800px) rotate3d(1, 1, 0, 90deg);
-      -webkit-animation-timing-function: ease-in;
-      animation-timing-function: ease-in;
-      opacity: 0;
-    }
-  
-    54% {
-      -webkit-transform: translateZ(-160px) rotate3d(1, 1, 0, 87deg);
-      transform: translateZ(-160px) rotate3d(1, 1, 0, 87deg);
-      -webkit-animation-timing-function: ease-in-out;
-      animation-timing-function: ease-in-out;
-      opacity: 1;
-    }
-  
-    100% {
-      -webkit-transform: translateZ(0) rotate3d(1, 1, 0, 0);
-      transform: translateZ(0) rotate3d(1, 1, 0, 0);
-      -webkit-animation-timing-function: ease-out;
-      animation-timing-function: ease-out;
-    }
-  }
-  
-  .profile {
-    -webkit-animation: slit-in-vertical 0.45s ease-out both;
-    animation: slit-in-vertical 0.45s ease-out both;
-  }
-  
-  .leaderboard {
-    -webkit-animation: slit-in-vertical 0.45s ease-out both;
-    animation: slit-in-vertical 0.45s ease-out both;
-  }
-  
-  .donation {
-    -webkit-animation: slit-in-vertical 0.45s ease-out both;
-    animation: slit-in-vertical 0.45s ease-out both;
-  }
-  
-  @-webkit-keyframes slit-in-vertical {
-    0% {
-      -webkit-transform: translateZ(-800px) rotateY(90deg);
-      transform: translateZ(-800px) rotateY(90deg);
-      opacity: 0;
-    }
-  
-    54% {
-      -webkit-transform: translateZ(-160px) rotateY(87deg);
-      transform: translateZ(-160px) rotateY(87deg);
-      opacity: 1;
-    }
-  
-    100% {
-      -webkit-transform: translateZ(0) rotateY(0);
-      transform: translateZ(0) rotateY(0);
-    }
-  }
-  
-  @keyframes slit-in-vertical {
-    0% {
-      -webkit-transform: translateZ(-800px) rotateY(90deg);
-      transform: translateZ(-800px) rotateY(90deg);
-      opacity: 0;
-    }
-  
-    54% {
-      -webkit-transform: translateZ(-160px) rotateY(87deg);
-      transform: translateZ(-160px) rotateY(87deg);
-      opacity: 1;
-    }
-  
-    100% {
-      -webkit-transform: translateZ(0) rotateY(0);
-      transform: translateZ(0) rotateY(0);
-    }
-  }
-  
-  .pageheaderGame {
-    -webkit-animation: tracking-in-expand 0.7s cubic-bezier(0.215, 0.610, 0.355, 1.000) both;
-    animation: tracking-in-expand 0.7s cubic-bezier(0.215, 0.610, 0.355, 1.000) both;
-  }
-  
-  .pageheader {
-    -webkit-animation: tracking-in-expand 0.7s cubic-bezier(0.215, 0.610, 0.355, 1.000) both;
-    animation: tracking-in-expand 0.7s cubic-bezier(0.215, 0.610, 0.355, 1.000) both;
-  }
-  
-  @-webkit-keyframes tracking-in-expand {
-    0% {
-      letter-spacing: -0.5em;
-      opacity: 0;
-    }
-  
-    40% {
-      opacity: 0.6;
-    }
-  
-    100% {
-      opacity: 1;
-    }
-  }
-  
-  @keyframes tracking-in-expand {
-    0% {
-      letter-spacing: -0.5em;
-      opacity: 0;
-    }
-  
-    40% {
-      opacity: 0.6;
-    }
-  
-    100% {
-      opacity: 1;
-    }
-  }
-  
-  .bg {
-    -webkit-animation: color-change-2x 6s linear infinite alternate both;
-    animation: color-change-2x 6s linear infinite alternate both;
-  }
-  
-  @-webkit-keyframes color-change-2x {
-    0% {
-      background: #1a1a1a;
-    }
-  
-    100% {
-      background: #001529;
-    }
-  }
-  
-  @keyframes color-change-2x {
-    0% {
-      background: #1a1a1a;
-    }
-  
-    100% {
-      background: #001529;
-    }
-  }
-  
-  @media screen and (max-width: 640px) {
-    .pageheaderGame {
-      font-size: 30px !important;
-    }
-  }
-  
-  @media screen and (max-width: 300px) {
-    .pageheaderGame {
-      font-size: 15px !important;
-    }
-  }
