+++ server/package.json
@@ -15,7 +15,8 @@
       "express": "^4.17.2",
       "graphql": "^15.5.0",
       "jsonwebtoken": "^8.5.1",
-      "mongoose": "^5.9.10"
+      "mongoose": "^5.9.10",
+      "stripe": "^12.8.0"
     },
     "devDependencies": {
       "nodemon": "^2.0.3"
