+++ server/schemas/resolvers.js
@@ -1,5 +1,6 @@
 const { User, Score } = require('../models');
 const { AuthenticationError } = require('apollo-server-express');
+const { signToken } = require('../utils/auth');
 
 const resolvers = {
     Query: {
