+++ server/schemas/typeDefs.js
@@ -1,3 +1,4 @@
+const { gql } = require('apollo-server-express');
 
 // template typeDefs, subject to change
 const typeDefs = `
